// src/models/seats-aero.model.ts
export type Route = {
  ID: string;
  OriginAirport: string;
  OriginRegion: string;
  DestinationAirport: string;
  DestinationRegion: string;
  NumDaysOut: number;
  Distance: number;
  Source: string;
};

export type AvailabilityData = {
  ID: string;
  RouteID: string;
  Route: Route;
  Date: string;
  ParsedDate: string;
  YAvailable: boolean;
  WAvailable: boolean;
  JAvailable: boolean;
  FAvailable: boolean;
  YMileageCost: string;
  WMileageCost: string;
  JMileageCost: string;
  FMileageCost: string;
  YDirectMileageCost: number;
  WDirectMileageCost: number;
  JDirectMileageCost: number;
  FDirectMileageCost: number;
  YRemainingSeats: number;
  WRemainingSeats: number;
  JRemainingSeats: number;
  FRemainingSeats: number;
  YDirectRemainingSeats: number;
  WDirectRemainingSeats: number;
  JDirectRemainingSeats: number;
  FDirectRemainingSeats: number;
  YAirlines: string;
  WAirlines: string;
  JAirlines: string;
  FAirlines: string;
  YDirectAirlines: string;
  WDirectAirlines: string;
  JDirectAirlines: string;
  FDirectAirlines: string;
  YDirect: boolean;
  WDirect: boolean;
  JDirect: boolean;
  FDirect: boolean;
  Source: string;
  CreatedAt: string;
  UpdatedAt: string;
  AvailabilityTrips: null;
};